move github to subfolder
